+++ src/datasource.ts
@@ -37,11 +37,6 @@ export class VictoriaLogsDatasource
 
   query(request: DataQueryRequest<Query>): Observable<DataQueryResponse> {
     const queries = request.targets.filter(q => q.expr).map((q) => {
-      // include time range in query if not already present
-      if (!/_time/.test(q.expr)) {
-        const timerange = `_time:[${request.range.from.toISOString()}, ${request.range.to.toISOString()}]`
-        q.expr = `${timerange} AND (${q.expr})`;
-      }
       return { ...q, maxLines: q.maxLines ?? this.maxLines }
     });